core: Fix version check
Previously it would report that Bash 4.0 and 4.1 were unsupported.

Since Bash always sets `BASH_VERSION`, it's not practical to write an
automated test unless different versions of the Bash binary are
available. I may add one, but since I've tested it manually on my
machine and this condition isn't likely to ever change, it's probably
not worth it.