template: Improve download_go_script_bash_tarball
Got rid of `GO_SCRIPT_BASH_VERSION_NUMBER`, since it was only used in
this function. Elided some of the `if` conditions into a single
`if/elif` chain. Made sure that the `mkdir -p` tries to create the full
parent of `GO_SCRIPT_BASH_CORE_DIR`, not just `GO_SCRIPTS_DIR`.