go-template: Use `cat` with `file://` URLs
This is because `wget` only supports HTTP, HTTPS, and FTP, as I
discovered when running the tests on Alpine Linux without cURL
installed.

Also, the BusyBox version of `tar` that comes with Alpine doesn't work
out of the box; the GNU version must be installed via `apk add tar`. Not
sure it's worth adding a check to `go-template` or not.