tests/template: Skip tests if programs are missing
Most of the test cases will now be skipped if any of the required
programs are missing, rather than failing and requiring the user to
investigate.

Whereas `skip_if_system_missing` will skip a test case if any of the
specified programs are missing, `skip_if_none_present_on_system` will
skip a test case only if all of the programs are missing. I'll move it
to `lib/bats/helpers` in a future commit.