.appveyor.yml: Put `/usr/bin` first in `PATH`
The previous commit failed with the same errors as before:

  https://ci.appveyor.com/project/mbland/go-script-bash/build/v1.6.0.3

Upon closer inspection, both of the failing tests executed `perl`.
Indeed, the Appveyor images have Perl installed separately from Git and
available in the PATH:

  https://www.appveyor.com/docs/build-environment/#perl

This Perl is most likely doing the Unix-to-native path translation. This
change attempts to fix the tests by ensuring the MSYS2 Perl that comes
with Git for Windows will get executed instead.